bug 修复

+++ src/main/java/com/aard/processor/serializable/ByteArraySerializable.java
@@ -69,7 +69,7 @@ public class ByteArraySerializable {
         int len = valueByteLength(value);
         byte[] bytes = new byte[len];
         HeapByteBufUtil.setByte(bytes, 0, TYPE);
-        HeapByteBufUtil.setByte(bytes, 1, value.length);
+        HeapByteBufUtil.setInt(bytes, 1, value.length);
         int index = 5;
         for (Byte val : value) {
             index = ByteSerializable.serializableValue(bytes, index, val);
