+++ src/main/java/com/aard/processor/serializable/LonArraySerializable.java
@@ -65,7 +65,7 @@ public class LonArraySerializable {
         int len = 5 + (value.length << 3);
         byte[] bytes = new byte[len];
         HeapByteBufUtil.setByte(bytes, 0, TYPE);
-        HeapByteBufUtil.setByte(bytes, 1, value.length);
+        HeapByteBufUtil.setInt(bytes, 1, value.length);
         int index = 5;
         for (long val : value) {
             index = LonSerializable.serializableValue(bytes, index, val);
