+++ src/main/java/com/aard/processor/serializable/StringArraySerializable.java
@@ -69,7 +69,7 @@ public class StringArraySerializable {
         int len = valueByteLength(value);
         byte[] bytes = new byte[len];
         HeapByteBufUtil.setByte(bytes, 0, TYPE);
-        HeapByteBufUtil.setByte(bytes, 1, value.length);
+        HeapByteBufUtil.setInt(bytes, 1, value.length);
         int index = 5;
         for (String val : value) {
             index = StringSerializable.serializableValue(bytes, index, val);
